+++ processor/src/operations/io_ops.rs
@@ -1,7 +1,4 @@
-use vm_core::{
-    WORD_SIZE,
-    mast::{BasicBlockNode, MastNodeExt},
-};
+use vm_core::WORD_SIZE;
 
 use super::{ExecutionError, Felt, Process};
 use crate::{Host, errors::ErrorContext};
@@ -40,7 +37,7 @@ impl Process {
     /// - Returns an error if the address is not aligned to a word boundary.
     pub(super) fn op_mloadw(
         &mut self,
-        error_ctx: &ErrorContext<'_, impl MastNodeExt>,
+        error_ctx: &impl ErrorContext,
     ) -> Result<(), ExecutionError> {
         // get the address from the stack and read the word from current memory context
         let mut word: [Felt; WORD_SIZE] = self
@@ -68,10 +65,7 @@ impl Process {
     ///   initialized to ZEROs, and thus, if the specified address has never been written to, the
     ///   ZERO element is returned.
     /// - The element retrieved from memory is pushed to the top of the stack.
-    pub(super) fn op_mload(
-        &mut self,
-        error_ctx: &ErrorContext<'_, BasicBlockNode>,
-    ) -> Result<(), ExecutionError> {
+    pub(super) fn op_mload(&mut self, error_ctx: &impl ErrorContext) -> Result<(), ExecutionError> {
         let element = self
             .chiplets
             .memory
@@ -97,7 +91,7 @@ impl Process {
     /// - Returns an error if the address is not aligned to a word boundary.
     pub(super) fn op_mstorew(
         &mut self,
-        error_ctx: &ErrorContext<'_, impl MastNodeExt>,
+        error_ctx: &impl ErrorContext,
     ) -> Result<(), ExecutionError> {
         // get the address from the stack and build the word to be saved from the stack values
         let addr = self.stack.get(0);
@@ -130,7 +124,7 @@ impl Process {
     /// Thus, the net result of the operation is that the stack is shifted left by one item.
     pub(super) fn op_mstore(
         &mut self,
-        error_ctx: &ErrorContext<'_, BasicBlockNode>,
+        error_ctx: &impl ErrorContext,
     ) -> Result<(), ExecutionError> {
         // get the address and the value from the stack
         let ctx = self.system.ctx();
@@ -164,7 +158,7 @@ impl Process {
     /// - Returns an error if the address is not aligned to a word boundary.
     pub(super) fn op_mstream(
         &mut self,
-        error_ctx: &ErrorContext<'_, impl MastNodeExt>,
+        error_ctx: &impl ErrorContext,
     ) -> Result<(), ExecutionError> {
         const MEM_ADDR_STACK_IDX: usize = 12;
 
@@ -222,7 +216,7 @@ impl Process {
     pub(super) fn op_pipe(
         &mut self,
         host: &mut impl Host,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
+        error_ctx: &impl ErrorContext,
     ) -> Result<(), ExecutionError> {
         const MEM_ADDR_STACK_IDX: usize = 12;
 
@@ -236,16 +230,16 @@ impl Process {
         let words = host
             .advice_provider_mut()
             .pop_stack_dword()
-            .map_err(|err| err.into_exec_err_at_clk(clk, err_ctx))?;
+            .map_err(|err| err.into_exec_err_at_clk(clk, error_ctx))?;
 
         // write the words memory
         self.chiplets
             .memory
-            .write_word(ctx, addr_first_word, clk, words[0], err_ctx)
+            .write_word(ctx, addr_first_word, clk, words[0], error_ctx)
             .map_err(ExecutionError::MemoryError)?;
         self.chiplets
             .memory
-            .write_word(ctx, addr_second_word, clk, words[1], err_ctx)
+            .write_word(ctx, addr_second_word, clk, words[1], error_ctx)
             .map_err(ExecutionError::MemoryError)?;
 
         // replace the elements on the stack with the word elements (in stack order)
@@ -279,7 +273,7 @@ impl Process {
     pub(super) fn op_advpop(
         &mut self,
         host: &mut impl Host,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
+        err_ctx: &impl ErrorContext,
     ) -> Result<(), ExecutionError> {
         let value = host
             .advice_provider_mut()
@@ -298,7 +292,7 @@ impl Process {
     pub(super) fn op_advpopw(
         &mut self,
         host: &mut impl Host,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
+        err_ctx: &impl ErrorContext,
     ) -> Result<(), ExecutionError> {
         let word = host
             .advice_provider_mut()
@@ -328,9 +322,7 @@ mod tests {
         super::{MIN_STACK_DEPTH, Operation},
         Felt, Host, Process,
     };
-    use crate::{
-        AdviceSource, ContextId, DefaultHost, ExecutionError, MemoryError, errors::ErrorContext,
-    };
+    use crate::{AdviceSource, ContextId, DefaultHost, ExecutionError, MemoryError};
 
     #[test]
     fn op_push() {
@@ -721,9 +713,9 @@ mod tests {
 
         // emulate reading and writing in the same clock cycle
         process.ensure_trace_capacity();
-        process.op_mload(&ErrorContext::default()).unwrap();
+        process.op_mload(&()).unwrap();
         assert_matches!(
-            process.op_mstore(&ErrorContext::default()),
+            process.op_mstore(&()),
             Err(ExecutionError::MemoryError(MemoryError::IllegalMemoryAccess {
                 ctx: _,
                 addr: _,
@@ -740,9 +732,9 @@ mod tests {
 
         // emulate reading and writing in the same clock cycle
         process.ensure_trace_capacity();
-        process.op_mstore(&ErrorContext::default()).unwrap();
+        process.op_mstore(&()).unwrap();
         assert_matches!(
-            process.op_mstore(&ErrorContext::default()),
+            process.op_mstore(&()),
             Err(ExecutionError::MemoryError(MemoryError::IllegalMemoryAccess {
                 ctx: _,
                 addr: _,
@@ -759,8 +751,8 @@ mod tests {
 
         // emulate reading in the same clock cycle
         process.ensure_trace_capacity();
-        process.op_mload(&ErrorContext::default()).unwrap();
-        process.op_mload(&ErrorContext::default()).unwrap();
+        process.op_mload(&()).unwrap();
+        process.op_mload(&()).unwrap();
     }
 
     // HELPER METHODS